pyproject.toml
[project]
name = "research-and-development-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.4.1",
    "langchain-mcp-adapters>=0.0.5",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
]
